refactoring Entitys

+++ app/src/main/java/com/example/habitstracker/data/net/entity/DoneDateEntity.kt
@@ -1,3 +1,3 @@
 package com.example.habitstracker.data.net.entity
 
-data class DoneDateEntity(val date: Int, val habit_uid: String)
+data class DoneDateEntity(val date: Long, val habit_uid: String)
